+++ CHANGELOG.md
@@ -7,6 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- Remove log function calls from the stack trace when calling `logDebug` or
+  `logError` (@MightyCreak)
+- Log functions used to log only on stderr, now `logDebug` logs on stdout and
+  `logError` logs on stderr (@MightyCreak)
+
 ## 0.8.0 - 2023-04-03
 
 ### Added